Выделил логику расчёта доступных ходов и атак
Данная логика была выделена из класса GameController в класс CharacterActionService

Fixes #20
